Refactor fetch-advisories and resolve-versions scripts; extract helper functions and add lastUpdatedTemp.json
